+++ tests/test_cli.py
@@ -1,0 +1,5 @@
+from subprocess import call
+
+
+def test_can_call_boca():
+    assert call(['boca']) == 0